Fixed marvel comic dto, updated ui of the lazy list for the comics and increased page size

+++ app/src/main/java/com/reringuy/marvelcharacterviewer/models/MarvelComic.kt
@@ -4,6 +4,6 @@ data class MarvelComic(
     val id: Int,
     val title: String,
     val description: String,
-    val issueNumber: Int,
+    val issueNumber: Float,
     val thumbnail: MarvelThumbnail
 )
